Create new app - 'mailings'
